+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 0.7.0
 
-* Added refresh button to right click folder menu
+* Added refresh button to right click folder menu & file menu
 
 ## 0.6.0